Refactor-file-structure
