+++ frontend/src/components/client/MenuBar.js
@@ -1,0 +1,38 @@
+import Meal from "./meals-section/Meal";
+import Tab from "../shared-components/Tab";
+import {useContext, useState} from 'react';
+import {MealsContext} from "../../Context/Context";
+
+
+export default function MenuBar() {
+
+
+  const days = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday'];
+  const { cartItems } = useContext(MealsContext)
+
+  const [activeDay, setActiveDay] = useState('');
+
+  const context = useContext(MealsContext)
+
+  const setDayHandler = (day) => {
+     setActiveDay(day);
+  }
+  
+
+  return (
+
+    <div className='flex flex-col p-5 lg:w-3/4 lg:h-2/3 sm:h-full sm:w-full justify-center self-center shadow-xl'>
+      <div className='flex w-full h-20 self-center justify-around'>
+        {days.map((day, index) => {
+          return (
+            <Tab key={index} day={day} role='day' active={activeDay} onDayChange={setDayHandler} />
+          )
+        })}
+      </div>
+      <div className='flex flex-col w-full h-full self-center overflow-y-scroll rounded-md'>
+        <Meal />
+      </div>
+      <button onClick={() => console.log(cartItems)} className='flex p-4 mt-5 duration-200 font-bold justify-center shadow-sm self-center items-center w-1/6 rounded-2xl text-white bg-green-400 hover:bg-green-500 focus:bg-green-700'>Confirm</button>
+    </div>
+  )
+}
