+++ frontend/src/components/operator/meals-section-operator/MealOperator.js
@@ -1,16 +1,14 @@
-import { useState } from 'react';
-import { MdArrowDropDown, MdArrowDropUp } from 'react-icons/md';
-import AddMeal from './AddMeal';
+import AddMeal from '../operatorView/OperatorAddMenu/AddMeal';
 import { useContext } from 'react';
-import { MealsContext } from '../../Context/Context';
+import { MealsContext } from '../../../Context/Context';
 import DishOperator from './DishOperator';
 
-export default function Meal(props) {
+export default function MealOperator(props) {
 
     const context = useContext(MealsContext)
 
     return (
-        <div className="flex h-auto text-lg justify-between flex-col items-center p-4 font-bold ">
+        <div className='flex h-auto text-lg justify-between flex-col items-center p-4 font-bold '>
             {context.show && <AddMeal />
             }
             <div className={`flex justify-center items-center w-full duration-200  hover:text-green-500 ${context.show ? 'text-green-500' : ''}`}>
